+++ src/app.py
@@ -8,7 +8,7 @@ import os
 from utils import models
 from utils.database import engine
 
-from routes.v1 import V1
+from routes import people, auth
 
 load_dotenv()
 
@@ -20,6 +20,9 @@ cors = CORSMiddleware(app, allow_origins=["*"])
 def health_check():
     return {"success": True}
 
-app.include_router(V1, prefix="/api/v1")
+app.include_router(auth.AUTH, prefix="/api/v1/auth", tags=["auth"])
+app.include_router(people.PEOPLE, prefix="/api/v1/people", tags=["people"])
+
+
 if __name__ == "__main__":
     uvicorn.run("app:app", host=os.environ.get("HOST"), port=int(os.environ.get("PORT")), reload=True)
